chore: agent multi added handle
